Replace stateManager method overriding with Proxy API
Replaced the unsustainable method overriding in _setupStateManager() to a much more robust Proxy API implementation.
We now create a proxified EVMEEI object that allows us to intercept putContractStorage and clearContractStorage without ever overriding them.
This guarantees us to be compliant with future changes in @ethereumjs/evm and still be able to capture storage clear/put operations.

Also, unrelated, this commit contains a fix to getTargetEvmVersion() in order to correctly get the target evmVersion for compilation.

Lastly, this commit contains a lint fix in DocRow.tsx