Added `name` query parameter handling
Added a `name` query parameter handling code to the `Filters.tsx` component. This handler will set a filter on the `Name` column of a table (either opcodes or precompiles) with the value of the query parameter.
In example, `evm.codes/precompiled?name=ec` will yield a search on the precompiles table for precompiles containing the sequence `ec` (ecrecover, ecpairing).
On the other hand, `evm.codes/?name=self` will yield a search on the opcodes table for opcodes containing the sequence `self` (SELFDESTRUCT, SELFBALANCE).