update-evm-storage-banner (#279)
Updates copy of the EVM Storage banner to reference new features. Removes the link to the original launch blog from June, given that blog is no longer representative of the current state of the product. Takes the remaining "Try it out" link and moves it in-line just to reduce the footprint of the banner a little now that we only have one link.

**To test**: Compare [prod](https://evm.codes) to [preview](https://evm-codes-git-12-05-update-evm-storage-banner-smlxl.vercel.app/?fork=shanghai).

Resolves SML-4297